Use 'accelerometer' crate types/traits
Provides some limited interop between accelerometers
